+++ go-controller/pkg/ovn/master.go
@@ -328,7 +328,7 @@ func (oc *DefaultNetworkController) syncGatewayLogicalNetwork(node *kapi.Node, l
 			return err
 		}
 		relevantHostIPs, err := util.MatchAllIPStringFamily(utilnet.IsIPv6(hostIfAddr.IP), sets.List(hostAddrs))
-		if err != nil && err != util.NoIPError {
+		if err != nil && err != util.ErrorNoIP {
 			return err
 		}
 		if err := oc.addPolicyBasedRoutes(node.Name, hostIfAddr.IP.String(), l3GatewayConfigIP, relevantHostIPs); err != nil {
